// forum-ui/src/app/api/models/login-result-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { OperationResultDto } from '../models/operation-result-dto';
export interface LoginResultDto {
  refreshToken?: string | null;
  refreshTokenExpires?: string | null;
  result?: OperationResultDto;
  token?: string | null;
}